+++ api/client/src/main/kotlin/su/plo/voice/api/client/config/ClientConfig.kt
@@ -79,6 +79,8 @@ interface ClientConfig {
 
         val panning: BooleanConfigEntry
 
+        val mutePlayerOnDirect: BooleanConfigEntry
+
         val cameraSoundListener: BooleanConfigEntry
 
         val exponentialVolumeSlider: BooleanConfigEntry